+++ pumaguard/classify.py
@@ -16,6 +16,9 @@ from pumaguard.traininghistory import TrainingHistory
 from pumaguard.utils import (
     initialize_tensorflow,
 )
+from pumaguard.models.pretrained import pre_trained_model
+from pumaguard.models.light import light_model
+from pumaguard.models.light_2 import light_model_2
 
 
 def get_duration(start_time: datetime.datetime,
@@ -32,108 +35,6 @@ def get_duration(start_time: datetime.datetime,
     """
     duration = end_time - start_time
     return duration / datetime.timedelta(microseconds=1) / 1e6
-
-
-def pre_trained_model(presets: Presets) -> keras.src.Model:
-    """
-    The pre-trained model (Xception).
-
-    Returns:
-        The model.
-    """
-    base_model = keras.applications.Xception(
-        weights='imagenet',
-        include_top=False,
-        input_shape=(*presets.image_dimensions, 3),
-    )
-
-    print(f'Number of layers in the base model: {len(base_model.layers)}')
-    print(f'shape of output layer: {base_model.layers[-1].output_shape}')
-
-    # We do not want to change the weights in the Xception model (imagenet
-    # weights are frozen)
-    base_model.trainable = False
-
-    # Average pooling takes the 2,048 outputs of the Xeption model and brings
-    # it into one output. The sigmoid layer makes sure that one output is
-    # between 0-1. We will train all parameters in these last two layers
-    return keras.Sequential([
-        base_model,
-        keras.layers.GlobalAveragePooling2D(),
-        keras.layers.Dense(1, activation='sigmoid'),
-    ])
-
-
-def light_model(presets: Presets) -> keras.src.Model:
-    """
-    Define the "light model" which is loosely based on the Xception model and
-    constructs a CNN.
-
-    Note, the light model does not run properly on a TPU runtime. The loss
-    function results in `nan` after only one epoch. It does work on GPU
-    runtimes though.
-    """
-    inputs = keras.Input(shape=(*presets.image_dimensions, 1))
-
-    # Entry block
-    x = keras.layers.Rescaling(1.0 / 255)(inputs)
-    x = keras.layers.Conv2D(128, 1, strides=2, padding="same")(x)
-    x = keras.layers.BatchNormalization()(x)
-    x = keras.layers.Activation("relu")(x)
-
-    previous_block_activation = x  # Set aside residual
-
-    for size in [256, 512, 728]:
-        x = keras.layers.Activation("relu")(x)
-        x = keras.layers.SeparableConv2D(size, 1, padding="same")(x)
-        x = keras.layers.BatchNormalization()(x)
-
-        x = keras.layers.Activation("relu")(x)
-        x = keras.layers.SeparableConv2D(size, 1, padding="same")(x)
-        x = keras.layers.BatchNormalization()(x)
-
-        x = keras.layers.MaxPooling2D(1, strides=2, padding="same")(x)
-
-        # Project residual
-        residual = keras.layers.Conv2D(size, 1, strides=2, padding="same")(
-            previous_block_activation
-        )
-        x = keras.layers.add([x, residual])  # Add back residual
-        previous_block_activation = x  # Set aside next residual
-
-    x = keras.layers.SeparableConv2D(1024, 1, padding="same")(x)
-    x = keras.layers.BatchNormalization()(x)
-    x = keras.layers.Activation("relu")(x)
-
-    x = keras.layers.GlobalAveragePooling2D()(x)
-
-    x = keras.layers.Dropout(0.1)(x)
-
-    outputs = keras.layers.Dense(1, activation=None)(x)
-
-    return keras.Model(inputs, outputs)
-
-
-def light_model_2(presets: Presets) -> keras.src.Model:
-    """
-    Another attempt at a light model.
-    """
-    return keras.Sequential([
-        keras.layers.Conv2D(
-            32,
-            (3, 3),
-            activation='relu',
-            input_shape=(*presets.image_dimensions, 1),
-        ),
-        keras.layers.MaxPooling2D(2, 2),
-        keras.layers.Conv2D(64, (3, 3), activation='relu'),
-        keras.layers.MaxPooling2D(2, 2),
-        keras.layers.Conv2D(128, (3, 3), activation='relu'),
-        keras.layers.MaxPooling2D(2, 2),
-        keras.layers.Flatten(),
-        keras.layers.Dense(512, activation='relu'),
-        keras.layers.Dense(1, activation='sigmoid'),
-    ])
 
 
 def create_model(presets: Presets,
